Updated code breaks

+++ StayWell/src/main/java/com/staywell/repository/CustomerDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.staywell.model.Customer;
 
-public interface CustomerDao extends JpaRepository<Customer, Integer>{
+public interface CustomerDao extends JpaRepository<Customer, Long>{
 
 	Optional<Customer> findByEmail(String email);
